% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dummies.R
\name{create.dummies}
\alias{create.dummies}
\title{Decode dummy variables}
\usage{
create.dummies(dataset, col, drop = TRUE)
}
\arguments{
\item{dataset}{the original dataset}

\item{col}{character the variable to decode}

\item{drop}{logical TRUE to drop the original column}
}
\description{
This function will create the dummy variables for the given col.
}

